+++ solutions/Virtual-Assistant/src/csharp/skills/emailskill/Dialogs/ForwardEmail/ForwardEmailDialog.cs
@@ -38,7 +38,7 @@ namespace EmailSkill
 
             var showEmail = new WaterfallStep[]
             {
-                IfClearContextStep,
+                PagingStep,
                 ShowEmails,
             };
