+++ solutions/Virtual-Assistant/src/csharp/skills/emailskill/Dialogs/ReplyEmail/ReplyEmailDialog.cs
@@ -34,7 +34,7 @@ namespace EmailSkill
 
             var showEmail = new WaterfallStep[]
             {
-                IfClearContextStep,
+                PagingStep,
                 ShowEmails,
             };
