+++ solutions/Virtual-Assistant/src/csharp/skills/emailskill/Dialogs/ShowEmail/ShowEmailDialog.cs
@@ -62,6 +62,7 @@ namespace EmailSkill
 
             var reshowEmail = new WaterfallStep[]
             {
+                PagingStep,
                 ShowEmailsWithoutEnd,
                 PromptToHandleMore,
                 HandleMore,
@@ -256,7 +257,7 @@ namespace EmailSkill
 
                 if (IsReadMoreIntent(topGeneralIntent, sc.Context.Activity.Text))
                 {
-                    return await sc.BeginDialogAsync(Actions.Show, skillOptions);
+                    return await sc.BeginDialogAsync(Actions.Reshow, skillOptions);
                 }
                 else if (topIntent == Email.Intent.Delete)
                 {
@@ -287,7 +288,7 @@ namespace EmailSkill
                 }
                 else if (topIntent == Email.Intent.None && (topGeneralIntent == General.Intent.Previous || topGeneralIntent == General.Intent.Next))
                 {
-                    return await sc.BeginDialogAsync(Actions.Show, skillOptions);
+                    return await sc.BeginDialogAsync(Actions.Reshow, skillOptions);
                 }
                 else
                 {
