+++ solutions/Virtual-Assistant/src/csharp/skills/emailskill/Startup.cs
@@ -114,7 +114,7 @@ namespace EmailSkill
                 // Typing Middleware (automatically shows typing when the bot is responding/working)
                 var typingMiddleware = new ShowTypingMiddleware();
                 options.Middleware.Add(typingMiddleware);
-                options.Middleware.Add(new SetLocaleMiddleware(defaultLocale ?? "en"));
+                options.Middleware.Add(new SetLocaleMiddleware(defaultLocale ?? "en-us"));
                 options.Middleware.Add(new AutoSaveStateMiddleware(userState, conversationState));
             });
         }
