+++ solutions/Virtual-Assistant/src/csharp/skills/emailskill/appsettings.json
@@ -19,5 +19,5 @@
     "readSize": "3",
     "displaySize": "3"
   },
-  "defaultLocale": "en"
+  "defaultLocale": "en-us"
 }
